requirements.txt
cryptography
pycryptodome
seaborn
numpy